Fix cpu demo crash on aarch64 platform (#252)
We try to fix #244 in this PR. The root cause of the crash is still not clear. After doing some experiments, we think that it isn't likely to be the bug of BladeDISC itself.

The crash could be re-produced on a graph with a single GEMM op using BladeDISC TF bridge.
Same GEMM op test is ok if we do the test in UT environments (not using TF bridge) or use BladeDISC torch bridge.
Codegen kernel runs successfully using BladeDISC TF bridge.
if we use the wheel file from ARM community instead of pip install tensorflow-aarch64, GEMM op test is ok using BladeDISC TF bridge.
Thus, we provide a workaround to fix #244 by replacing the wheel package.