[RAL] refactor metadata file: remove protobuf dependency. (#620)
* [RAL] refactor metadata file: remove protobuf dependency.

The metadata file is generated during compilation and is used to aid the
execution of the compiled binary. it includes the (large) weight values
extracted from the original AI model. Protobuf does not support file larger
than 2GB, leading to the failure of compiling some large models.

This CL replaces the protobuf-based solution with a customized storage
schema.

* fix

* fix

* fix

* fix

* fix

* refine